+++ audio-streaming-client-java/src/main/java/com/baidu/acu/pie/demo/JavaDemo.java
@@ -39,13 +39,14 @@ public class JavaDemo {
                 .serverPort(80)
                 .appName("simple demo")
                 .sleepRatio(1)
-                .product(AsrProduct.CUSTOMER_SERVICE_FINANCE);
+                .product(AsrProduct.CUSTOMER_SERVICE_FINANCE)
+                .productId("1906");
 
         return AsrClientFactory.buildClient(asrConfig);
     }
 
     public void recognizeFile() {
-        String audioFilePath = "testaudio/1.wav";
+        String audioFilePath = "testaudio/bj8k.wav";
         AsrClient asrClient = createAsrClient();
         List<RecognitionResult> results = asrClient.syncRecognize(Paths.get(audioFilePath));
 
@@ -54,6 +55,7 @@ public class JavaDemo {
 
         for (RecognitionResult result : results) {
             System.out.println(String.format(AsrConfig.TITLE_FORMAT,
+                    "file_name",
                     "serial_num",
                     "err_no",
                     "err_message",
@@ -61,6 +63,7 @@ public class JavaDemo {
                     "end_time",
                     "result"));
             System.out.println(String.format(AsrConfig.TITLE_FORMAT,
+                    audioFilePath,
                     result.getSerialNum(),
                     result.getErrorCode(),
                     result.getErrorMessage(),
@@ -75,19 +78,26 @@ public class JavaDemo {
      * 使用长音频来模拟人对着麦克风不断说话的情况
      */
     public void asyncRecognition() {
-        String longAudioFilePath = "testaudio/1.wav";
+        String longAudioFilePath = "testaudio/300s.wav";
         AsrClient asrClient = createAsrClient();
         CountDownLatch finishLatch = new CountDownLatch(1);
-        StreamObserver<AudioFragmentRequest> sender = asrClient.asyncRecognize(it -> {
-            System.out.println(
-                    Instant.now().toString() + "\t" + Thread.currentThread().getId() + " receive fragment: " + it);
-        }, finishLatch);
 
+        StreamObserver<AudioFragmentRequest> sender = asrClient.asyncRecognize(
+                // 这个回调函数用来处理每一个fragment的识别结果
+                it -> {
+                    System.out.println(
+                            Instant.now().toString() + "\t" + Thread.currentThread().getId()
+                                    + " receive fragment: " + it);
+                }, finishLatch);
+
+        // 这里从文件中得到一个InputStream，实际场景下，也可以从麦克风或者其他音频源来得到InputStream
         try (InputStream audioStream = Files.newInputStream(Paths.get(longAudioFilePath))) {
             byte[] data = new byte[asrClient.getFragmentSize()];
             int readSize;
 
             System.out.println(Instant.now().toString() + "\t" + Thread.currentThread().getId() + " start to send");
+
+            // 使用 sender.onNext 方法，将 InputStream 中的数据不断地发送到 asr 后端，发送的最小单位是 AudioFragment
             while ((readSize = audioStream.read(data)) != -1) {
                 System.out.println(Instant.now().toString() + "\t"
                         + Thread.currentThread().getId() + " send fragment");
@@ -95,6 +105,7 @@ public class JavaDemo {
                         .setAudioData(ByteString.copyFrom(data, 0, readSize))
                         .build());
                 // 主动休眠一段时间，来模拟人说话场景下的音频产生速率
+                // 在对接麦克风等设备的时候，可以去掉这个 sleep
                 Thread.sleep(20);
             }
             System.out.println(Instant.now().toString() + "\t" + Thread.currentThread().getId() + " send finish");